+++ CHANGELOG.md
@@ -27,6 +27,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - `traceIDRatioSampler` (given by `TraceIDRatioBased(float64)`) now uses the rightmost bits for sampling decisions,
   fixing random sampling when using ID generators like `xray.IDGenerator`
   and increasing parity with other language implementations. (#3557)
+- The OTLP exporters for traces and metrics will print the final retryable error message when attempts to retry time out. (#3514)
 
 ### Deprecated
 
@@ -34,8 +35,6 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Changed
 
-- OTLP exporter for traces and metrics will print the final retryable error message
-  when attempts to retry time out. (#3514)
 
 ## [1.11.2/0.34.0] 2022-12-05